Migrated Dropdown component to TypeScript
